requirements.txt
Telethon==1.29.3
python-dotenv==1.0.0
